UBUNTU: [Packaging] Add versioned dependency for linux-base
BugLink: http://bugs.launchpad.net/bugs/1767133

We Depends on linux-base because of linux-update-symlinks and
linux-check-removal. Those have been introduced in a version later
than 4.0ubuntu1, which shipped on Xenial.

The unversioned Depends makes some upgrades to Bionic and
installations of linux-hwe-edge to break because those tools are
not found during the configure phase.

Follow this change from the master branch.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>