UBUNTU: [Packaging] Rename debian.azure to debian.azure-4.15
There are tools (at least cranky fix) which find a potential master's
debian directory by parsing the branch name of the master. So whenever
a branch is a master for another branch, the specific debian directory
must be named debian.<xxx>. Where xxx is the branch name which has a
potential '-next' postfix stripped.

Ignore: yes
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>