UBUNTU: SAUCE: ubuntu/sgx: Add module alias for ACPI device INT0E0C
BugLink: https://bugs.launchpad.net/bugs/1881338

Based on version LD_1.22 and:

https://github.com/intel/SGXDataCenterAttestationPrimitives/pull/115/commits/8390589ed10ac2ec431d6d607a5ac71d10bc9c72

The INT0E0C device denotes SGX support in the DSDT ACPI table. Add
this as an alias to the SGX driver. This allows the module to be
auto-probed when the module is build with the kernel (rather than
DKMS) and the entry is present in the machine's DSDT.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>