CIFS: Do not miss cancelled OPEN responses
BugLink: https://bugs.launchpad.net/bugs/1918714

When an OPEN command is cancelled we mark a mid as
cancelled and let the demultiplex thread process it
by closing an open handle. The problem is there is
a race between a system call thread and the demultiplex
thread and there may be a situation when the mid has
been already processed before it is set as cancelled.

Fix this by processing cancelled requests when mids
are being destroyed which means that there is only
one thread referencing a particular mid. Also set
mids as cancelled unconditionally on their state.

Cc: Stable <stable@vger.kernel.org>
Tested-by: Frank Sorenson <sorenson@redhat.com>
Reviewed-by: Ronnie Sahlberg <lsahlber@redhat.com>
Signed-off-by: Pavel Shilovsky <pshilov@microsoft.com>
Signed-off-by: Steve French <stfrench@microsoft.com>
(backported from commit 7b71843fa7028475b052107664cbe120156a2cfc)
[ fs/cifs/transport.c has changed significantly, I pieced in the
intended changes to the old function DeleteMidQEntry() ]
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>

v2 - ran scripts/checkpatch.pl and corrected formatting errors.
Acked-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>