Drivers: hv: vmbus: Implement Direct Mode for stimer0
BugLink: https://bugs.launchpad.net/bugs/1882623

The 2016 version of Hyper-V offers the option to operate the guest VM
per-vcpu stimer's in Direct Mode, which means the timer interupts on its
own vector rather than queueing a VMbus message. Direct Mode reduces
timer processing overhead in both the hypervisor and the guest, and
avoids having timer interrupts pollute the VMbus interrupt stream for
the synthetic NIC and storage.  This patch enables Direct Mode by
default on stimer0 when running on a version of Hyper-V that supports
it.

In prep for coming support of Hyper-V on ARM64, the arch independent
portion of the code contains calls to routines that will be populated
on ARM64 but are not needed and do nothing on x86.

Signed-off-by: Michael Kelley <mikelley@microsoft.com>
Signed-off-by: K. Y. Srinivasan <kys@microsoft.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(backported from commit 248e742a396e7f00b283f1c56e14b1bef6e3ec56)
[gpiccoli: context adjustments, specially due to the lack of the
following commits in our kernel:
- commit 51d4e5daa328 ("x86/irq: Count Hyper-V reenlightenment interrupts")
- commit 5a4858032217 ("x86/hyper-v: move hyperv.h out of uapi")
Note that the hyperv.h uapi file changed to hyperv-tlfs.h not in uapi.]
Signed-off-by: Guilherme G. Piccoli <gpiccoli@canonical.com>
Acked-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Kelsey Skunberg <kelsey.skunberg@canonical.com>